Introduce a SideBarFactory interface.
This interface is implemented on the fly by Ninject to create the required
instance of the SideBar, currently hard-bound to ConcertsSideBar.  It
removes part of the Abstract Factory pattern boiler plate, but it
(obviously) doesn't provide the logic required to bind a specific side bar
based on the requested action's controller.