Resolve the sidebar by the controller name.
Add the requesting controller to the route data, and then pull it out of
the route data to construct an appropriate side bar object for the page.

The SideBarFactory is currently combining the concerns of a
ConcertsSideBarFactory, an EmptySideBarFactory, and a
SideBarFactoryFactory (deciding which factory to use).  This needs to be
rectified, probably once I've got the persistence sorted.