Map Concert Description property.
The Description property on a Concert was not being mapped between the domain model and the view model.  Additionally, the value of the property was still hard-coded in the Concert model, so this class has been updated as well to provide a public setter.  This means that, for now at least, Concert domain models are no longer immutable.