Persist changes to the edited Concert.
If the view model is all ok, then the Concert can be persisted to the database.  RavenDB tracks changes to the model retrieved via a call to Load() or Query(), so this object must have its state mutated.  As such, I have introduced methods onto the Concert domain model to enable mutation of its state.