Use separate view models for the REST API.
Create a new separate Concert view model for the REST API to contain the appropriate Web API attributes, and to be better represent Concerts on the wire.  This should help pin down the problems with the quick actions on the Concerts' list page (no concert ID is being presented to the UI).

Bring in AutoMapper from NuGet to handle the object-object mappings.  The manual approach in the MapperExtensions class was clunky but worked ok for just one mapping; now there are two mappings, it makes sense to bring in a more robust framework for handling this work.