Implement quick publishing of concerts.
This is a fairly big change.  The API view model has been restructured to better represent what is displayed on-screen (e.g., single DateTime field for the concert date and start time).  The Knockout ViewModels for the Concerts feature have also been restructured to reflect this.  I'm not happy with the way this is done at the moment, and it seems like I need to refresh (and improve) my Knockout knowledge: the UI items bound to the IsPublished flag are not updating when the flag's value changes.

This change also enables camel-casing in JSON.NET, so there's no need to do case transitions in the Knockout ViewModels anymore.