Fix up Concerts list page.
Another big check-in, but this completes the initial work on Concerts.  Refactor the Knockout Concert ViewModel to be fully observable, and the Concert list ViewModel to bind to the Quick Add form: this means the quick add action can pull the data out of the ViewModel rather than having to query the form fields directly.  Publishing a concert now auto-updates the page as expected too.

Move away from HTML Helpers for the Quick Add form: the code is clearer using plain HTML and there's no model binding going on here that might be assisted by their use.

Complete the implementation of the POST action on the Concerts API controller.  A successful POST request should, in addition to returning a 201 Created, return the URI of the new resource in the Location header.  I have also opted to return the created object over the wire, as appears to be the best practice.  This has two somewhat unfortunate side effects:

  - Duplication of the new resource's URI: this is returned in the content of the response as well as in the Location header, as I have an Href property on the wire representation of a Concert.  This should be revisited, as it may be possible to pull out the URI for the client from the response.
  - The received view model must first be converted to a model for persisting in Raven, and then the saved model must be converted back to a view model for returning to the client.  This is to ensure the new resource's ID is set correctly in the response content.

Additionally, there seems to be a lot of set-up required for correct operation of the Web API during unit testing a POST request.  The complex logic and context required to correctly instantiate an ApiController will probably need pulling out into an extension method or something when I get to writing another controller.