Move date format logic to the View.
How to display a given date object is more of a View concern than a ViewModel concern, so move the application of the date format to this location.  The application of the format is delegated to a helper function, now moved to (and imported from) a new file, commonFunctions.js.