Enable client-side validation on quick-add form.
Crikey, this was way tougher than it needed to have been.  A few things to note:

  - The unobtrusive validation stuff was interfering with my manual validation. Something to add to the backlog (port to unobtrusive validation)?
  - The version of jQuery Validation I was using (1.8) did not support the HTML 5 input types, so I've updated to the latest version of both jQuery and jQuery Validation.
  - I switched the format of the date saved by the date picker from yyyy-mm-dd to dd/mm/yyyy.  This then required the same change to be made to the min attribute on the quick add form.

There also seems to be a bunch of whitespace changes that I can't be bothered to undo at this point.

The next thing to do is to get the validation messages to display more nicely so they don't fuck with the layout of the quick add form as is currently the case.