Apply the Abstract Factory pattern better.
Instead of switching on controller name to decide what type of SideBar to create, inject a different instance of ISideBarFactory depending on what type is requesting it.

This requires a DefaultSideBarFactory and DefaultSideBar to be used when a specific one is not available, to avoid binding exceptions.  The DefaultSideBar just returns an empty Enumerable of SideBarSections (i.e., is equivalent to no side-bar).