Fix Ninject bindings for the Side Bar.
As before, only inject the ConcertsSideBar when a Concerts page is requested.  The change made is to use Bind() rather than Rebind(), which is advised against (http://stackoverflow.com/questions/9209687/issue-rebind-ninject-while-dynamically-change-the-connection-string/9209932#9209932).  The way to fix the operation of Bind() over what I had before is to explicitly exclude all implementations of ISideBarFactory.