Allow each View to determine how wide it should be.
This involves moving the choice of spanX class to each view (currently
only the Concerts and home page views are updated), and moving the Side
Bar to each view. I don't much like this latter part of things, as th Side
Bar is most likely to be displayed on all views in a "section" (e.g.
concerts), rather than just some.