Fix up Admin Area visual design.
Bring it into line with the main site by copy-pasting the HTML from the
main site. There's a lot of duplicatin in this area now, which indicates
that the Layouts perhaps could be consolidated.

Update a few of the Admin Area Views to make the pages render more
correctly (i.e. remove the page-header as they are already displayed via
the Admin Area Layout).