Add CGO logo to navbar.
Display a small version of the CGO logo in the site's navbar in the
brand area. Shrink the brand text down from "Cambridge Graduate
Orchestra" to "CGO", and add the logo to the left of the brand text so
it is nicely padded within the navbar and displays in line with the left
border of the body content.