Fix the navbar brand's padding.
Pad it to 15px on both sides to ensure there's sufficient padding
between the brand and the first navbar menu item. This is the default
Bootstrap value for padding of this item, and so blends in nicely. It
also means the nasty margin-left hack I'd put in on the brand image
(logo) can be removed!