Add the background image at different resolutions.
Include the appropriate background image by screen width, using a set of
media queries.

I'm not 100% sure this is correct, or that I've got the images quite
right: at the moment, the images are sized to the minimum dimensions
required (e.g. 992px wide for screen-md), but I think this should be
changed to the maximum dimensions (i.e. the minimum dimensions for the
next size up, so 1200px wide for screen-md) to avoid artefacts appearing
in the background image.