Fix up nav bars.
Replace the per-item drop-down lists on the main site navbar with a single
button to take you to the admin site.

Add drop-down lists to each item in the admin site for create, and
edit/delete.

Add a "Back to Main Site" link in the Admin area, and make the brand area
link to the admin home.